+++ src/css-rules/no-px-unit/index.ts
@@ -1,0 +1,3 @@
+import noPxUnitRule from './rule-definition.js'
+
+export default noPxUnitRule